Simplify snippets
