+++ docs/action-results.md
@@ -2,7 +2,7 @@
 
 MVC Action results support allow composition handlers to set custom response results for specific scenarios, like for example, handling bad requests or validation error thoat would nornmally require throwing an exception. Setting a custom action result is done by using the `SetActionResult()` `HttpRequest` extension method:
 
-<!-- snippet: net-core-3x-action-results -->
+<!-- snippet: action-results -->
 <a id='snippet-net-core-3x-action-results'></a>
 ```cs
 public class UseSetActionResultHandler : ICompositionRequestsHandler
@@ -31,7 +31,7 @@ public class UseSetActionResultHandler : ICompositionRequestsHandler
 
 Using MVC action results require enabling output formatters support:
 
-<!-- snippet: net-core-3x-action-results-required-config -->
+<!-- snippet: action-results-required-config -->
 <a id='snippet-net-core-3x-action-results-required-config'></a>
 ```cs
 services.AddViewModelComposition(options =>
