+++ docs/custom-http-status-codes.md
@@ -2,7 +2,7 @@
 
 The response status code can be set in requests handlers and it'll be honored by the composition pipeline. To set a custom response status code the following snippet can be used:
 
-<!-- snippet: net-core-3x-sample-handler-with-custom-status-code -->
+<!-- snippet: sample-handler-with-custom-status-code -->
 <a id='snippet-net-core-3x-sample-handler-with-custom-status-code'></a>
 ```cs
 public class SampleHandlerWithCustomStatusCode : ICompositionRequestsHandler
