+++ docs/output-formatters-serialization.md
@@ -4,7 +4,7 @@ _Available starting with v1.9.0_
 
 Enabling output formatters support is a metter of:
 
-<!-- snippet: net-core-3x-use-output-formatters -->
+<!-- snippet: use-output-formatters -->
 <a id='snippet-net-core-3x-use-output-formatters'></a>
 ```cs
 public void ConfigureServices(IServiceCollection services)
@@ -30,7 +30,7 @@ The required steps are:
 
 The above configuration uses the new `System.Text.Json` serializer as the default json serializer to format json responses. The `System.Text.Json` serializer does not support serializing `dynamic` objects. It's possible to plug-in the Newtonsoft Json.Net serializer as output formatter by adding a package reference to the `Microsoft.AspNetCore.Mvc.NewtonsoftJson` package, and using the following configuration:
 
-<!-- snippet: net-core-3x-use-newtonsoft-output-formatters -->
+<!-- snippet: use-newtonsoft-output-formatters -->
 <a id='snippet-net-core-3x-use-newtonsoft-output-formatters'></a>
 ```cs
 public void ConfigureServices(IServiceCollection services)
