+++ docs/response-serialization-casing.md
@@ -8,7 +8,7 @@ _Available starting with v1.8.0_
 
 Default response serialization is camel casing. It's possible to configure a different default response serialization casing when configuring ServiceComposer:
 
-<!-- snippet: net-core-3x-default-casing -->
+<!-- snippet: default-casing -->
 <a id='snippet-net-core-3x-default-casing'></a>
 ```cs
 public void ConfigureServices(IServiceCollection services)
