+++ src/Snippets/ModelBinding/ConfigureAppForModelBinding.cs
@@ -5,7 +5,7 @@ namespace Snippets.NetCore3x.ModelBinding
 {
     public class ConfigureAppForModelBinding
     {
-        // begin-snippet: net-core-3x-model-binding-add-controllers
+        // begin-snippet: model-binding-add-controllers
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddViewModelComposition();
