+++ src/Snippets/SampleHandler/SampleHandler.cs
@@ -7,7 +7,7 @@ using ServiceComposer.AspNetCore;
 
 namespace Snippets.NetCore3x.SampleHandler
 {
-    // begin-snippet: net-core-3x-sample-handler-with-authorization
+    // begin-snippet: sample-handler-with-authorization
     public class SampleHandlerWithAuthorization : ICompositionRequestsHandler
     {
         [Authorize]
@@ -19,7 +19,7 @@ namespace Snippets.NetCore3x.SampleHandler
     }
     // end-snippet
 
-    // begin-snippet: net-core-3x-sample-handler-with-custom-status-code
+    // begin-snippet: sample-handler-with-custom-status-code
     public class SampleHandlerWithCustomStatusCode : ICompositionRequestsHandler
     {
         [HttpGet("/sample/{id}")]
