+++ src/Snippets/Serialization/Startup.cs
@@ -1,17 +1,21 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using Newtonsoft.Json;
 using ServiceComposer.AspNetCore;
 
-namespace Snippets.NetCore3x.DefaultCasing
+namespace Snippets.NetCore3x.Serialization
 {
     public class Startup
     {
-        // begin-snippet: net-core-3x-default-casing
+        // begin-snippet: custom-serialization-settings
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddRouting();
             services.AddViewModelComposition(options =>
             {
-                options.ResponseSerialization.DefaultResponseCasing = ResponseCasing.PascalCase;
+                options.ResponseSerialization.UseCustomJsonSerializerSettings(request =>
+                {
+                    return new JsonSerializerSettings();
+                });
             });
         }
         // end-snippet
