+++ src/Snippets/Snippets.csproj
